+++ pomoapp-frontend/src/App.tsx
@@ -1,12 +1,15 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
-import Login from './components/Login';
-import Dashboard from './components/Dashboard';
+import Login from './pages/Login';
+import Dashboard from './pages/Dashboard';
 
 const App = () => {
   return (
     <BrowserRouter>
       <Routes>
+        {/* ログインはフルスクリーンで単体表示 */}
         <Route path="/" element={<Login />} />
+
+        {/* ダッシュボード用ルート */}
         <Route path="/dashboard" element={<Dashboard />} />
       </Routes>
     </BrowserRouter>
