+++ pomoapp-frontend/src/components/Header.tsx
@@ -1,0 +1,13 @@
+// src/components/Header.tsx
+import React from 'react';
+import './Header.css';
+
+const Header = () => {
+  return (
+    <header className="header">
+      <h1 className="logo">Pomodoro App</h1>
+    </header>
+  );
+};
+
+export default Header;
