+++ pomoapp-frontend/src/components/Sidebar.css
@@ -1,0 +1,22 @@
+/* src/components/Sidebar.css */
+.sidebar {
+  width: 200px;
+  background-color: #000;
+  color: #0f0;
+  height: 100vh;
+  padding: 1rem;
+  box-shadow: 2px 0 5px #0f0;
+}
+
+.sidebar ul {
+  list-style: none;
+  padding: 0;
+}
+
+.sidebar li {
+  margin-bottom: 1rem;
+  cursor: pointer;
+  text-shadow: 0 0 5px #0f0;
+  font-size: 1.3rem;
+  margin-bottom: 1.5rem;
+}
